requirements.txt
openai
yt-dlp
pysrt
openai-whisper
